fix table collation being blank sometimes
new issue on the table info screen, look at the collation dropdown for a table, then switch tables. sometimes the collation is there, sometimes not.

This fixes that, the table encoding was not being cleared on reset (i.e. when you switch tables)
So it was looking up the collation for the previous table.