Possible active window cycling fix for issue #1666
This commit enables the 'Visible At Launch' checkbox option in MainWindow.xib
I was experiencing an issue where Command+backtick would not cycle
through 2 or more open Sequel Ace windows unless they each contained
2 or more tabs. For some reason this completely fixes my issue.
I sort of stumbled on this fix by accident as I'm an
iOS developer and not very familiar with NSWindow/NSWindowController.
Please vet this and maybe use it as a reference for a proper fix as I feel
there is probably a more intentional solution.
For reference I am on an M1 running Ventura 13.2.1 and do NOT use
Stage Manager or Spaces.